Fix : An unitialized data access.
DETAILS :

We detected sevral problems with running comparefiles with pdfwrite with
enabling floating point interrupts (patch IM866).
This patch fixes an unitialized data access, or inaccurate type casts.
These problems may relate to Bug 688379
"pdfwrite: an indeterministic behavior with Bug687660a.ps",
rather we could not get a proof for that
due to instability of the effect in the old code.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6180 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevpsft.c
@@ -399,7 +399,7 @@ write_mtx(stream *s, gs_font_type42 *pfont, const gs_type42_mtx_t *pmtx,
 {
     uint num_metrics = pmtx->numMetrics;
     uint len = num_metrics * 4;
-    double factor = pfont->data.unitsPerEm * (wmode ? -1 : 1);
+    double factor = (double)pfont->data.unitsPerEm * (wmode ? -1 : 1);
     float sbw[4];
     uint i;
